development_process/requirements.txt
Flask==2.2.2
pymysql==1.0.2
requests==2.28.1
beautifulsoup4==4.11.1
